+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. Version number will be decided later -->
 - feat: Add `utils/mergeRef` helper
+- feat: Add prop `wrapperRef` to `TableWrapper`
 - `useLaggyState` changes:
   - feat: Allow setting `customDelay` to `false` to emulate `useState()`
   - fix: Only set initial `thenState` once